+++ packages/network/lib/middlewares.dart
@@ -1,3 +1,2 @@
 export 'src/middlewares/default_errors.dart';
 export 'src/middlewares/encode_json_body.dart';
-export 'src/middlewares/throw_on_status_codes.dart';