+++ packages/network/lib/network.dart
@@ -4,5 +4,3 @@ export 'src/response.dart';
 export 'src/settings.dart';
 export 'src/network.dart';
 export 'src/middleware.dart';
-// TODO(Serge) remove in 1.0.0
-export 'src/hooks.dart';