+++ packages/network/lib/src/exception.dart
@@ -1,10 +1,10 @@
 import 'package:network/src/request.dart';
 import 'package:network/src/response.dart';
 
-class NetworkException<T extends BinaryResponse> implements Exception {
+class NetworkException implements Exception {
   NetworkException(this.response);
 
-  final T response;
+  final Response response;
 
   Request get request => response?.request;
 
@@ -12,10 +12,4 @@ class NetworkException<T extends BinaryResponse> implements Exception {
 
   @override
   String toString() => '$runtimeType{code: $code}';
-}
-
-class NetworkUnavailableException implements Exception {
-  NetworkUnavailableException(this.request);
-
-  final Request request;
 }