+++ packages/network/lib/src/middlewares/default_errors.dart
@@ -2,67 +2,67 @@ import 'package:network/src/exception.dart';
 import 'package:network/src/middleware.dart';
 import 'package:network/src/response.dart';
 
-Middleware defaultErrors() => Middleware(
-      onError: (Object error) {
-        if (error is NetworkException &&
-            !(error is NetworkUnavailableException)) {
-          switch (error.code) {
+Middleware defaultErrors({
+  int moreThan = 400,
+}) =>
+    Middleware(
+      onResponse: (res) {
+        if (res.statusCode >= moreThan) {
+          switch (res.statusCode) {
             case 400:
-              return BadRequestException(error.response);
+              throw BadRequestException(res);
             case 401:
-              return UnauthorizedException(error.response);
+              throw UnauthorizedException(res);
             case 402:
-              return PaymentRequired(error.response);
+              throw PaymentRequired(res);
             case 403:
-              return ForbiddenException(error.response);
+              throw ForbiddenException(res);
             case 404:
-              return NotFoundException(error.response);
+              throw NotFoundException(res);
             case 405:
-              return MethodNotAllowed(error.response);
+              throw MethodNotAllowed(res);
             case 500:
-              return InternalServerException(error.response);
+              throw InternalServerException(res);
             default:
-              return error;
+              throw NetworkException(res);
           }
         } else {
-          return error;
+          return res;
         }
       },
     );
 
 /// HTTP 400
-class BadRequestException<T extends BinaryResponse>
-    extends NetworkException<T> {
-  BadRequestException(T response) : super(response);
+class BadRequestException extends NetworkException {
+  BadRequestException(Response response) : super(response);
 }
 
 /// HTTP 401
-class UnauthorizedException<T extends BinaryResponse> extends NetworkException {
-  UnauthorizedException(T response) : super(response);
+class UnauthorizedException extends NetworkException {
+  UnauthorizedException(Response response) : super(response);
 }
 
 /// HTTP 402
-class PaymentRequired<T extends BinaryResponse> extends NetworkException<T> {
-  PaymentRequired(T response) : super(response);
+class PaymentRequired extends NetworkException {
+  PaymentRequired(Response response) : super(response);
 }
 
 /// HTTP 403
-class ForbiddenException<T extends BinaryResponse> extends NetworkException {
-  ForbiddenException(T response) : super(response);
+class ForbiddenException extends NetworkException {
+  ForbiddenException(Response response) : super(response);
 }
 
 /// HTTP 404
-class NotFoundException<T extends BinaryResponse> extends NetworkException<T> {
-  NotFoundException(T response) : super(response);
+class NotFoundException extends NetworkException {
+  NotFoundException(Response response) : super(response);
 }
 
 // HTTP 405
-class MethodNotAllowed<T extends BinaryResponse> extends NetworkException<T> {
-  MethodNotAllowed(T response) : super(response);
+class MethodNotAllowed extends NetworkException {
+  MethodNotAllowed(Response response) : super(response);
 }
 
 /// HTTP 500
-class InternalServerException<T extends BinaryResponse>
-    extends NetworkException<T> {
-  InternalServerException(T response) : super(response);
+class InternalServerException extends NetworkException {
+  InternalServerException(Response response) : super(response);
 }